# ⚡️ Simpler, faster `ENVELOPE` and `address` parser