* Envelope can *NOT* be `NIL`.
* Using aliased rules for envelope and address parts
* Using QUOTED_rev2 regexp for IMAP4rev2 and UTF8=ACCEPT address strings
* remove extraneous `envelope_data` method

Benchmarks:
```
Calculating -------------------------------------
                          v0.4.5-6-gbc32ddb2-dirty        0.4.4
      bodystructure_mixed_boundary       3.379k           3.109k i/s
                  10.152k times in       3.004058s        3.265234s
rfc3501_8_example_3_FETCH_ENVELOPE       5.355k           4.646k i/s
                  15.924k times in       2.973921s        3.427366s

Comparison:
                   bodystructure_mixed_boundary
          v0.4.5-6-gbc32ddb2-dirty:      3379.4 i/s
                             0.4.4:      3109.1 i/s - 1.09x  slower

             rfc3501_8_example_3_FETCH_ENVELOPE
          v0.4.5-6-gbc32ddb2-dirty:      5354.5 i/s
                             0.4.4:      4646.1 i/s - 1.15x  slower
```